# api/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Генерация отчета по продажам за месяц.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Generate Geport
  version: "1.0"
paths:
...
  /sales/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет запись о продаже по ID
      parameters:
      - description: ID записи о продаже
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Succesfully deleted
          schema:
            type: string
        "400":
          description: Неверный ID
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Удаление записи о продаже
      tags:
      - sales
...
securityDefinitions:
  BearerAuth:
    description: 'Введите JWT токен следующим образом: "Bearer {токен}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"